detection/requirements.txt
tensorflow==2.12.0
numpy==1.23.0
opencv-python==4.7.0
Pillow==9.5.0
